+++ db-service/test/cqn4sql/tupleExpansion.test.js
@@ -1,7 +1,7 @@
 'use strict'
 
 const cqn4sql = require('../../lib/cqn4sql')
-const cds = require('@sap/cds/lib')
+const cds = require('@sap/cds')
 const { expect } = cds.test
 // TODO test for unsupported comparison ops
 describe('Structural comparison', () => {