+++ eslint.config.js
@@ -1,8 +1,6 @@
 'use strict'
 
 const eslint_js = require('@eslint/js')
-const eslint_ts = require('@typescript-eslint/eslint-plugin')
-const typescript_parser = require('@typescript-eslint/parser')
 const globals = require('globals')
 
 module.exports = [
@@ -34,23 +32,6 @@ module.exports = [
       'no-extra-semi': 'off',
       'no-unused-vars': ['warn', { argsIgnorePattern: 'lazy' }],
       'no-console': 'error',
-    },
-  },
-  {
-    files: ['**/*.ts'],
-    plugins: {
-      '@typescript-eslint': eslint_ts,
-    },
-    languageOptions: {
-      parser: typescript_parser,
-      globals: {
-        ...globals.node,
-      },
-    },
-    rules: {
-      ...eslint_ts.configs.recommended.rules,
-      '@typescript-eslint/no-unused-vars': ['warn', { argsIgnorePattern: '^_' }],
-      '@typescript-eslint/no-explicit-any': ['warn'],
     },
   },
   {